Extracted the CopyButton from AssetDetailsBox
